Refactor HeapAttackIT to run on Serverless
